extracting Model from the View
