+++ unity-renderer/Assets/Features/CameraReel/_Common/CameraReelModel.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8a600df978854c2a995479f4a8cda7ca
+timeCreated: 1690800703
